power: avoid implicit casts of int64_t to double
As suggested by clang-tidy, avoid implicit cast of long and int64_t to
double by making the cast explicit.

While arguably stylistically slightly better, this change is effectively
a no-op. However, a more useful change would require significant design
work.

BUG=b:258053283
TEST=cros_run_unit_tests --board=brya --packages power_manager

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4028171
Commit-Queue: Alan Green <avg@chromium.org>
Tested-by: Alan Green <avg@chromium.org>

+++ power_manager/tools/backlight_tool.cc
@@ -163,13 +163,15 @@ class Converter {
 
   // Converts a brightness level to a linear percent in [0.0, 100.0].
   double LevelToLinearPercent(int64_t level) {
-    return level * 100.0 / backlight_.GetMaxBrightnessLevel();
+    return static_cast<double>(level) * 100.0 /
+           static_cast<double>(backlight_.GetMaxBrightnessLevel());
   }
 
   // Converts a linear percent in [0.0, 100.0] to a brightness level.
   int64_t LinearPercentToLevel(double percent) {
     return static_cast<int64_t>(roundl(
-        ClampPercent(percent) * backlight_.GetMaxBrightnessLevel() / 100.0));
+        ClampPercent(percent) *
+        static_cast<double>(backlight_.GetMaxBrightnessLevel()) / 100.0));
   }
 
   // Returns the initial brightness level requested by |controller_|.
